requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
channels
psycopg2
Pillow
python-dotenv
asgiref
drf-yasg